Explicit casting in normalisation

+++ Grid/qcd/observables/exp_scalar.h
@@ -69,7 +69,7 @@ class ExpScalarLogger : public HmcObservable<typename Impl::Field> {
                           GridSerialRNG &sRNG,
                           GridParallelRNG &pRNG) {
 
-    double e = sum(trace(exp(Pars.a*U))) / U._grid->gSites();;
+    double e = sum(trace(exp(Pars.a*U))) / (double)U._grid->gSites();;
 
     int def_prec = std::cout.precision();
